Revert "Cabal 2.0 internal-lib reorg"
This reverts commit 9646fb92cf6b2213cdd17499eec19769d73860ec.
