+++ Network/DNS/Base32Hex.hs
