+++ Network/DNS/LookupRaw.hs
@@ -19,8 +19,8 @@ import Network.DNS.IO
 import Network.DNS.Imports hiding (lookup)
 import Network.DNS.Memo
 import Network.DNS.Transport
+import Network.DNS.Types
 import Network.DNS.Types.Internal
-import Network.DNS.Types.Resolver
 
 -- $setup
 -- >>> :set -XOverloadedStrings
