+++ Network/DNS/StateBinary.hs
@@ -65,7 +65,7 @@ import qualified Data.Map as M
 import Data.Semigroup as Sem
 
 import Network.DNS.Imports
-import Network.DNS.Types.Internal
+import Network.DNS.Types
 
 ----------------------------------------------------------------
