+++ appveyor.yml
@@ -21,8 +21,8 @@ install:
  - "set PATH=C:\\msys64\\mingw64\\bin;C:\\msys64\\usr\\bin;C:\\ghc\\ghc-%GHCVER%\\bin;%PATH%"
  - "cabal --version"
  - "ghc --version"
- - "cabal %CABOPTS% new-update"
- - "cabal %CABOPTS% new-install hspec-discover"
+ - "cabal %CABOPTS% update -vverbose+nowrap"
+ - "cabal %CABOPTS% install hspec-discover"
 
 build: off
 deploy: off
@@ -31,5 +31,8 @@ test_script:
  - cabal %CABOPTS% new-build -j1 -vnormal+nowrap --dry all
  - ps: Push-AppveyorArtifact dist-newstyle\cache\plan.json
  - cabal %CABOPTS% new-build -j1 -vnormal+nowrap all
- - cabal %CABOPTS% new-test  -j1 -vnormal+nowrap all
+ # On Windows, doctests don't yet work due to missing cbits
+ # dependency, so we run just the 'network' and 'spec' tests.
+ - cabal %CABOPTS% new-test  -j1 -vnormal+nowrap network
+ - cabal %CABOPTS% new-test  -j1 -vnormal+nowrap spec
  - ps: ls dns*.log -recurse | %{ Push-AppveyorArtifact $_}
