+++ dns.cabal
@@ -10,46 +10,15 @@ Description:
   in pure Haskell.
 Category:               Network
 Cabal-Version:          >= 2.0
-Build-Type:             Simple
+Build-Type:             Custom
 Extra-Source-Files:     Changelog.md
                         cbits/dns.c
 Tested-With:            GHC == 8.2.2
                       , GHC == 8.4.4
                       , GHC == 8.6.5
 
-Library dns-internal
-  Default-Language:     Haskell2010
-  GHC-Options:          -Wall
-  hs-source-dirs:       Internal
-  Exposed-Modules:      Network.DNS.Imports
-                        Network.DNS.Types.Internal
-                        Network.DNS.Types.Resolver
-                        Network.DNS.Resolver.Internal
-                        Network.DNS.Decode.Parsers
-                        Network.DNS.Decode.Internal
-                        Network.DNS.Encode.Builders
-                        Network.DNS.Encode.Internal
-                        Network.DNS.StateBinary
-                        Network.DNS.Memo
-                        Network.DNS.Base32Hex
-  build-depends:        base
-                      , array
-                      , async
-                      , attoparsec
-                      , auto-update
-                      , base16-bytestring
-                      , base64-bytestring
-                      , bytestring
-                      , containers
-                      , cryptonite
-                      , hourglass
-                      , iproute
-                      , mtl
-                      , network
-                      , psqueues
-  if os(windows)
-    C-Sources:        cbits/dns.c
-    Extra-Libraries:  iphlpapi
+Custom-Setup
+  Setup-Depends:        base, Cabal, cabal-doctest >=1.0.6 && <1.1
 
 Library
   Default-Language:     Haskell2010
@@ -61,11 +30,21 @@ Library
                         Network.DNS.Utils
                         Network.DNS.Types
                         Network.DNS.Decode
+                        Network.DNS.Decode.Internal
                         Network.DNS.Encode
+                        Network.DNS.Encode.Internal
                         Network.DNS.IO
-  Other-Modules:        Network.DNS.Transport
-  Build-Depends:        dns-internal
-                      , base >= 4 && < 5
+  Other-Modules:        Network.DNS.Base32Hex
+                        Network.DNS.Decode.Parsers
+                        Network.DNS.Encode.Builders
+                        Network.DNS.Imports
+                        Network.DNS.Memo
+                        Network.DNS.StateBinary
+                        Network.DNS.Transport
+                        Network.DNS.Types.Internal
+  if impl(ghc < 8)
+    Build-Depends:      semigroups
+  Build-Depends:        base >= 4 && < 5
                       , array
                       , async
                       , attoparsec
@@ -80,8 +59,12 @@ Library
                       , mtl
                       , network >= 2.3
                       , psqueues
+  if os(windows)
+    Build-Depends:    split
+    C-Sources:        cbits/dns.c
+    Extra-Libraries:  iphlpapi
 
-Test-Suite network-tests
+Test-Suite network
   Type:                 exitcode-stdio-1.0
   Default-Language:     Haskell2010
   Hs-Source-Dirs:       test2
@@ -90,12 +73,11 @@ Test-Suite network-tests
   Other-Modules:        LookupSpec
                         IOSpec
   Build-Depends:        dns
-                      , dns-internal
                       , base
                       , hspec
                       , network
 
-Test-Suite spec-tests
+Test-Suite spec
   Type:                 exitcode-stdio-1.0
   Default-Language:     Haskell2010
   Hs-Source-Dirs:       test
@@ -105,7 +87,6 @@ Test-Suite spec-tests
                         DecodeSpec
                         RoundTripSpec
   Build-Depends:        dns
-                      , dns-internal
                       , QuickCheck >= 2.9
                       , base
                       , bytestring
@@ -113,11 +94,15 @@ Test-Suite spec-tests
                       , iproute >= 1.3.2
                       , word8
 
+-- The doctests don't presently work on Windows, due to an unresolved
+-- cbits/dns.c dependency.  Documentation on how to link custom C-code
+-- via cabal-doctest is not provided.
+--
 Test-Suite doctests
   Type:                 exitcode-stdio-1.0
   Default-Language:     Haskell2010
   Hs-Source-Dirs:       test2
-  Ghc-Options:          -Wall -threaded
+  Ghc-Options:          -Wall
   Main-Is:              doctests.hs
   Build-Depends:        base
                       , doctest
